[wrapper] Split channel.rs into builder_channel and replay_channel; colocate CircuitWire impls
Moves each CircuitWire impl next to the channel that uses it, and splits the
mixed channel.rs into one file per channel.

- channel.rs → builder_channel.rs holds IronSpartanBuilderChannel + the
  BuilderWire CircuitWire impl (formerly in circuit_elem.rs).
- New replay_channel.rs holds ReplayChannel + the WitnessGenWire
  CircuitWire impl (formerly in circuit_elem.rs).
- zk_wrapped_channel.rs absorbs WrappedWire and NoopBuilder (formerly in
  circuit_elem.rs).
- circuit_elem.rs is reduced to the CircuitWire trait, the CircuitElem
  enum + its arithmetic/FieldOps impls, and the OwnedOrRef helper.

The wrapper/mod.rs module declarations and re-exports are updated to
reflect the new layout (builder_channel, replay_channel, zk_wrapped_channel
are all `pub mod`; IronSpartanBuilderChannel, ReplayChannel, and
ZKWrappedVerifierChannel are still re-exported at the wrapper module
root). Test imports point at the new module paths.

No behavior changes — purely a code reorganization. All tests pass.
